==> 6-questionnaires.ipynb <==
COLOR1 = '#21427f'
COLOR2 = '#FF7518'

import pickle
import os
import numpy as np
import matplotlib.pyplot as plt
import pandas as pd

nasa_tlx = pd.read_csv('data/nasa_tlx.csv', index_col=0)
display(nasa_tlx)

#Column daw and vr are two bars to plot, 2Dse and 3Dse are the two error bars

# Plotting the NASA-TLX scores
fig, ax = plt.subplots(figsize=(6, 3.2))
index = np.arange(nasa_tlx.shape[0])
bar_width = 0.35
opacity = 1

rects1 = plt.bar(index -bar_width/2, nasa_tlx['DAW'], bar_width, alpha=opacity, color=COLOR1, label='DAW', edgecolor='black', zorder=10)
rects2 = plt.bar(index + bar_width/2, nasa_tlx['VR'], bar_width, alpha=opacity, color=COLOR2, label='VR', edgecolor='black', zorder=10)

plt.errorbar(index - bar_width/2, nasa_tlx['DAW'], yerr=nasa_tlx['2D SE'], fmt='none', ecolor='black', capsize=4, zorder=10)
plt.errorbar(index + bar_width/2, nasa_tlx['VR'], yerr=nasa_tlx['3D SE'], fmt='none', ecolor='black', capsize=4, zorder=10)

# Statistical significance asterisks
asterisks = {'Performance':'*', "Temporal demand" : '* *'}

for i,row in enumerate(nasa_tlx.index):
    # print('row:', row)
    for key, val in asterisks.items():
        if row.lower() == key.lower():
            print('Adding "%s" to column %s at position %d' % (val, key, i))

            ypos = 15 #max(nasa_tlx['DAW'][key], nasa_tlx['VR'][key]) + 1
            plt.text(i, ypos, val, ha='center', va='bottom', color='black', fontsize=13)



# plt.xlabel('Subject')
plt.ylabel('Self-report score')
# plt.title('NASA-TLX Scores')

plt.xticks(index, nasa_tlx.index.str.strip().str.replace(' ','\n'), rotation=0)
# plt.xticks(index, nasa_tlx.index.str.strip(), rotation=15, ha='right')
topylim = 17
plt.ylim(0, topylim)
plt.yticks(np.arange(0, topylim+1, 4))
# ygrid
plt.grid(axis='y', linestyle='--', alpha=0.6, zorder=0)

plt.legend()
plt.tight_layout()

if os.path.exists('plots/questionnaires') is False:
    os.makedirs('plots/questionnaires')
plt.savefig('plots/questionnaires/nasa_tlx.pdf', bbox_inches='tight')
plt.savefig('plots/questionnaires/nasa_tlx.png', bbox_inches='tight')




ad_hoc = pd.read_csv('data/ad_hoc.csv', index_col=0)
display(ad_hoc)



# Plotting the Ad Hoc scores

fig, ax = plt.subplots(figsize=(12, 5))
index = np.arange(ad_hoc.shape[0])
bar_width = 0.35
opacity = 1
spacing = 0

rects1 = plt.bar(index -bar_width/2, ad_hoc['DAW'], bar_width, alpha=opacity, color=COLOR1, label='DAW', edgecolor='black', zorder=10)
rects2 = plt.bar(index + bar_width/2, ad_hoc['VR'], bar_width, alpha=opacity, color=COLOR2, label='VR', edgecolor='black', zorder=10)

plt.errorbar(index - bar_width/2, ad_hoc['DAW'], yerr=ad_hoc['DAW SE'], fmt='none', ecolor='black', capsize=4, zorder=10)
plt.errorbar(index + bar_width/2, ad_hoc['VR'], yerr=ad_hoc['VR SE'], fmt='none', ecolor='black', capsize=4, zorder=10)

# Statistical significance asterisks
asterisks = {'Standing':'* * *',
             'Satisfaction':'* *',
             'enjoyment':'* * *',
             'Helpful':'*',
             'Effort Reward':'*',
             'Expressiveness':'* *',
             'Absorption':'* *',
             'creativity':'*',
             'attention':'*'}

for i,row in enumerate(ad_hoc.index):
    # print('row:', row)
    for key, val in asterisks.items():
        if row.lower() == key.lower():
            print('Adding "%s" to column %s at position %d' % (val, key, i))

            ypos = max(ad_hoc['DAW'][row]+ad_hoc['DAW SE'][row], ad_hoc['VR'][row]+ad_hoc['VR SE'][row]) + 0
            plt.text(i*(1+spacing), ypos, val, ha='center', verticalalignment='center', va='bottom', color='black', fontsize=12)



# plt.xlabel('Subject')
plt.ylabel('Self-report score')
# plt.title('Ad Hoc Scores')

# plt.xticks(index, ad_hoc.index.str.strip(), rotation=15, ha='right')
plt.xticks(index, ad_hoc.index.str.strip(), rotation=15)
# plt.xticks(index, ad_hoc.index.str.strip())
# plt.ylim(0, 16)
plt.yticks(np.arange(0, 11, 1))
# ygrid
plt.grid(axis='y', linestyle='--', alpha=0.6, zorder=0)

plt.legend()
plt.tight_layout()

if os.path.exists('plots/questionnaires') is False:
    os.makedirs('plots/questionnaires')

plt.savefig('plots/questionnaires/ad_hoc.pdf', bbox_inches='tight')
plt.savefig('plots/questionnaires/ad_hoc.png', bbox_inches='tight')

# Behavioral data now (which needs mean and se computation to be done here)

# Load the data
behavioral_raw = pd.read_csv('data/behavioral.csv', index_col=[0,1])
# Sort by index 0 and 1, where 0 is "P" plus an integer, and 1 is a string
display(behavioral_raw)

behavioral_statistics = pd.DataFrame(index = behavioral_raw.columns, columns=['DAW', 'VR', 'DAW SE', 'VR SE'])

# Compute mean and standard error
for col in behavioral_raw.columns:
    # first, compute the mean of column col for all players (index[0]) and condition (index[1]) DAW
    behavioral_statistics.loc[col, 'DAW'] = behavioral_raw.loc[(slice(None), 'DAW'), col].mean()
    behavioral_statistics.loc[col, 'DAW SE'] = behavioral_raw.loc[(slice(None), 'DAW'), col].std() / np.sqrt(behavioral_raw.loc[(slice(None), 'DAW'), col].shape[0])
    
    behavioral_statistics.loc[col, 'VR'] = behavioral_raw.loc[(slice(None), 'VR'), col].mean()
    behavioral_statistics.loc[col, 'VR SE'] = behavioral_raw.loc[(slice(None), 'VR'), col].std() / np.sqrt(behavioral_raw.loc[(slice(None), 'VR'), col].shape[0])

index_order = ['play/stop', 'select track','draw trajectory', 'clear automation', 'change timeslider' ]
behavioral_statistics = behavioral_statistics.loc[index_order]
# first in caps lock in index
behavioral_statistics.index = behavioral_statistics.index.str.capitalize()

display(behavioral_statistics)

fig, ax = plt.subplots(figsize=(6, 5))
index = np.arange(behavioral_statistics.shape[0])
bar_width = 0.35
opacity = 1

# del spacing

rects1 = plt.bar(index -bar_width/2, behavioral_statistics['DAW'], bar_width, alpha=opacity, color=COLOR1, label='DAW', edgecolor='black', zorder=10)
rects2 = plt.bar(index + bar_width/2, behavioral_statistics['VR'], bar_width, alpha=opacity, color=COLOR2, label='VR', edgecolor='black', zorder=10)

plt.errorbar(index - bar_width/2, behavioral_statistics['DAW'], yerr=behavioral_statistics['DAW SE'], fmt='none', ecolor='black', capsize=4, zorder=10)
plt.errorbar(index + bar_width/2, behavioral_statistics['VR'], yerr=behavioral_statistics['VR SE'], fmt='none', ecolor='black', capsize=4, zorder=10)

# Statistical significance asterisks
asterisks = {'change timeslider':'*'}

for i,row in enumerate(behavioral_statistics.index):
    # print('row:', row)
    for key, val in asterisks.items():
        if row.lower() == key.lower():
            print('Adding "%s" to column %s at position %d' % (val, key, i))

            ypos = max(behavioral_statistics['DAW'][row]+behavioral_statistics['DAW SE'][row], behavioral_statistics['VR'][row]+behavioral_statistics['VR SE'][row]) + 0
            plt.text(i, ypos, val, ha='center', verticalalignment='center', va='bottom', color='black', fontsize=12)

plt.ylabel('Score')

plt.xticks(index, behavioral_statistics.index.str.strip().str.replace(' ','\n'), size=9) #, rotation=15)
plt.yticks(np.arange(0, 24, 2))
plt.grid(axis='y', linestyle='--', alpha=0.6, zorder=0)

plt.legend()

plt.savefig('plots/questionnaires/behavioral.pdf', bbox_inches='tight')
plt.savefig('plots/questionnaires/behavioral.png', bbox_inches='tight')
